+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2026-04-21
+
+### Added
+
+- Nextcloud groups can now also be assigned based on hitobito event participations of the user
+
+### Changed
+
+- Updated dependencies
+
 ## [1.1.3] - 2026-03-17
 
 ### Changed